+++ R/rope.R
@@ -1,6 +1,6 @@
-#' Region of Practical Equivalence (ROPE)
+#' Region of Practical Equivalence (ROPE) Analysis
 #'
-#' Compute the proportion of the HDI (default to the `89%` HDI) of a posterior
+#' Compute the proportion of the CI (default to the 95% ETI) of a posterior
 #' distribution that lies within a region of practical equivalence.
 #'
 #' @param x Vector representing a posterior distribution. Can also be a
@@ -25,50 +25,53 @@
 #' proportion of HDI, to use for the percentage in ROPE.
 #' @param ci_method The type of interval to use to quantify the percentage in
 #' ROPE. Can be 'HDI' (default) or 'ETI'. See [`ci()`].
+#' @param complement Should the probabilities above/below the ROPE (the
+#' _complementary_ probabilities) be returned as well? See
+#' [equivalence_test()] as well.
 #'
 #' @inheritParams hdi
 #'
 #' @inheritSection hdi Model components
 #'
 #' @section ROPE:
-#' Statistically, the probability of a posterior distribution of being
-#' different from 0 does not make much sense (the probability of a single value
-#' null hypothesis in a continuous distribution is 0). Therefore, the idea
+#'
+#' Statistically, the probability of a posterior distribution of being different
+#' from 0 does not make much sense (the probability of a single value null
+#' hypothesis in a continuous distribution is 0). Therefore, the idea
 #' underlining ROPE is to let the user define an area around the null value
 #' enclosing values that are *equivalent to the null* value for practical
 #' purposes (_Kruschke 2010, 2011, 2014_).
 #'
-#' Kruschke (2018) suggests that such null value could be set, by default,
-#' to the -0.1 to 0.1 range of a standardized parameter (negligible effect
-#' size according to Cohen, 1988). This could be generalized: For instance,
-#' for linear models, the ROPE could be set as `0 +/- .1 * sd(y)`.
-#' This ROPE range can be automatically computed for models using the
-#' [`rope_range()`] function.
+#' Kruschke (2018) suggests that such null value could be set, by default, to
+#' the -0.1 to 0.1 range of a standardized parameter (negligible effect size
+#' according to Cohen, 1988). This could be generalized: For instance, for
+#' linear models, the ROPE could be set as `0 +/- .1 * sd(y)`. This ROPE range
+#' can be automatically computed for models using the [`rope_range()`] function.
 #'
-#' Kruschke (2010, 2011, 2014) suggests using the proportion of  the `95%`
-#' (or `89%`, considered more stable) [HDI][hdi] that falls within the
-#' ROPE as an index for "null-hypothesis" testing (as understood under the
-#' Bayesian framework, see [`equivalence_test()`]).
+#' Kruschke (2010, 2011, 2014) suggests using the proportion of [HDI][hdi] that
+#' falls within the ROPE as an index for "null-hypothesis" testing (as
+#' understood under the Bayesian framework, see [`equivalence_test()`]).
 #'
 #' @section Sensitivity to parameter's scale:
-#' It is important to consider the unit (i.e., the scale) of the predictors
-#' when using an index based on the ROPE, as the correct interpretation of the
-#' ROPE as representing a region of practical equivalence to zero is dependent
-#' on the scale of the predictors. Indeed, the percentage in ROPE depend on
-#' the unit of its parameter. In other words, as the ROPE represents a fixed
-#' portion of the response's scale, its proximity with a coefficient depends
-#' on the scale of the coefficient itself.
+#'
+#' It is important to consider the unit (i.e., the scale) of the predictors when
+#' using an index based on the ROPE, as the correct interpretation of the ROPE
+#' as representing a region of practical equivalence to zero is dependent on the
+#' scale of the predictors. Indeed, the percentage in ROPE depend on the unit of
+#' its parameter. In other words, as the ROPE represents a fixed portion of the
+#' response's scale, its proximity with a coefficient depends on the scale of
+#' the coefficient itself.
 #'
 #' @section Multicollinearity - Non-independent covariates:
+#'
 #' When parameters show strong correlations, i.e. when covariates are not
-#' independent, the joint parameter distributions may shift towards or
-#' away from the ROPE. Collinearity invalidates ROPE and hypothesis
-#' testing based on univariate marginals, as the probabilities are conditional
-#' on independence. Most problematic are parameters that only have partial
-#' overlap with the ROPE region. In case of collinearity, the (joint) distributions
-#' of these parameters may either get an increased or decreased ROPE, which
-#' means that inferences based on `rope()` are inappropriate
-#' (_Kruschke 2014, 340f_).
+#' independent, the joint parameter distributions may shift towards or away from
+#' the ROPE. Collinearity invalidates ROPE and hypothesis testing based on
+#' univariate marginals, as the probabilities are conditional on independence.
+#' Most problematic are parameters that only have partial overlap with the ROPE
+#' region. In case of collinearity, the (joint) distributions of these
+#' parameters may either get an increased or decreased ROPE, which means that
+#' inferences based on `rope()` are inappropriate (_Kruschke 2014, 340f_).
 #'
 #' `rope()` performs a simple check for pairwise correlations between
 #' parameters, but as there can be collinearity between more than two variables,
@@ -175,20 +178,32 @@ rope.default <- function(x, ...) {
 
 #' @rdname rope
 #' @export
-rope.numeric <- function(x,
-                         range = "default",
-                         ci = 0.95,
-                         ci_method = "ETI",
-                         verbose = TRUE,
-                         ...) {
+rope.numeric <- function(
+  x,
+  range = "default",
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  verbose = TRUE,
+  ...
+) {
   if (all(range == "default")) {
     range <- c(-0.1, 0.1)
   } else if (!all(is.numeric(range)) || length(range) != 2) {
-    insight::format_error("`range` should be 'default' or a vector of 2 numeric values (e.g., c(-0.1, 0.1)).")
+    insight::format_error(
+      "`range` should be 'default' or a vector of 2 numeric values (e.g., c(-0.1, 0.1))."
+    )
   }
 
   rope_values <- lapply(ci, function(i) {
-    .rope(x, range = range, ci = i, ci_method = ci_method, verbose = verbose)
+    .rope(
+      x,
+      range = range,
+      ci = i,
+      ci_method = ci_method,
+      complement = complement,
+      verbose = verbose
+    )
   })
 
   # "do.call(rbind)" does not bind attribute values together
@@ -196,7 +211,11 @@ rope.numeric <- function(x,
 
   out <- do.call(rbind, rope_values)
   if (nrow(out) > 1) {
-    out$ROPE_Percentage <- as.numeric(out$ROPE_Percentage)
+    iv <- intersect(
+      colnames(out),
+      c("ROPE_Percentage", "Superiority_Percentage", "Inferiority_Percentage")
+    )
+    out[iv] <- lapply(out[iv], as.numeric)
   }
 
   # Attributes
@@ -216,13 +235,16 @@ rope.numeric <- function(x,
 
 
 #' @export
-rope.get_predicted <- function(x,
-                               range = "default",
-                               ci = 0.95,
-                               ci_method = "ETI",
-                               use_iterations = FALSE,
-                               verbose = TRUE,
-                               ...) {
+rope.get_predicted <- function(
+  x,
+  range = "default",
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  use_iterations = FALSE,
+  verbose = TRUE,
+  ...
+) {
   if (isTRUE(use_iterations)) {
     if ("iterations" %in% names(attributes(x))) {
       out <- rope(
@@ -230,6 +252,7 @@ rope.get_predicted <- function(x,
         range = range,
         ci = ci,
         ci_method = ci_method,
+        complement = complement,
         verbose = verbose,
         ...
       )
@@ -243,6 +266,7 @@ rope.get_predicted <- function(x,
       range = range,
       ci = ci,
       ci_method = ci_method,
+      complement = complement,
       verbose = verbose,
       ...
     )
@@ -254,13 +278,16 @@ rope.get_predicted <- function(x,
 #' @export
 #' @rdname rope
 #' @inheritParams p_direction
-rope.data.frame <- function(x,
-                            range = "default",
-                            ci = 0.95,
-                            ci_method = "ETI",
-                            rvar_col = NULL,
-                            verbose = TRUE,
-                            ...) {
+rope.data.frame <- function(
+  x,
+  range = "default",
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  rvar_col = NULL,
+  verbose = TRUE,
+  ...
+) {
   obj_name <- insight::safe_deparse_symbol(substitute(x))
 
   x_rvar <- .possibly_extract_rvar_col(x, rvar_col)
@@ -276,7 +303,7 @@ rope.data.frame <- function(x,
     return(.append_datagrid(out, x))
   }
 
-  out <- .prepare_rope_df(x, range, ci, ci_method, verbose)
+  out <- .prepare_rope_df(x, range, ci, ci_method, complement, verbose)
   HDI_area_attributes <- insight::compact_list(out$HDI_area)
   dat <- data.frame(
     Parameter = rep(names(HDI_area_attributes), each = length(ci)),
@@ -294,17 +321,21 @@ rope.data.frame <- function(x,
 
 
 #' @export
-rope.draws <- function(x,
-                       range = "default",
-                       ci = 0.95,
-                       ci_method = "ETI",
-                       verbose = TRUE,
-                       ...) {
+rope.draws <- function(
+  x,
+  range = "default",
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  verbose = TRUE,
+  ...
+) {
   rope(
     .posterior_draws_to_df(x),
     range = range,
     ci = ci,
     ci_method = ci_method,
+    complement = complement,
     verbose = verbose,
     ...
   )
@@ -315,14 +346,25 @@ rope.rvar <- rope.draws
 
 
 #' @export
-rope.emmGrid <- function(x,
-                         range = "default",
-                         ci = 0.95,
-                         ci_method = "ETI",
-                         verbose = TRUE,
-                         ...) {
-  xdf <- insight::get_parameters(x)
-  dat <- rope(xdf, range = range, ci = ci, ci_method = ci_method, verbose = verbose, ...)
+rope.emmGrid <- function(
+  x,
+  range = "default",
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  verbose = TRUE,
+  ...
+) {
+  xdf <- insight::get_parameters(x, verbose = verbose)
+  dat <- rope(
+    xdf,
+    range = range,
+    ci = ci,
+    ci_method = ci_method,
+    complement = complement,
+    verbose = verbose,
+    ...
+  )
   dat <- .append_datagrid(dat, x)
   attr(dat, "object_name") <- insight::safe_deparse_symbol(substitute(x))
   dat
@@ -333,18 +375,22 @@ rope.emm_list <- rope.emmGrid
 
 
 #' @export
-rope.slopes <- function(x,
-                        range = "default",
-                        ci = 0.95,
-                        ci_method = "ETI",
-                        verbose = TRUE,
-                        ...) {
+rope.slopes <- function(
+  x,
+  range = "default",
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  verbose = TRUE,
+  ...
+) {
   xrvar <- .get_marginaleffects_draws(x)
   dat <- rope(
     xrvar,
     range = range,
     ci = ci,
     ci_method = ci_method,
+    complement = complement,
     verbose = verbose,
     ...
   )
@@ -361,20 +407,24 @@ rope.predictions <- rope.slopes
 
 
 #' @export
-rope.BFBayesFactor <- function(x,
-                               range = "default",
-                               ci = 0.95,
-                               ci_method = "ETI",
-                               verbose = TRUE,
-                               ...) {
+rope.BFBayesFactor <- function(
+  x,
+  range = "default",
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  verbose = TRUE,
+  ...
+) {
   if (all(range == "default")) {
     range <- rope_range(x, verbose = verbose)
   }
   out <- rope(
-    insight::get_parameters(x),
+    insight::get_parameters(x, verbose = verbose),
     range = range,
     ci = ci,
     ci_method = ci_method,
+    complement = complement,
     verbose = verbose,
     ...
   )
@@ -387,18 +437,22 @@ rope.bamlss <- rope.BFBayesFactor
 
 
 #' @export
-rope.MCMCglmm <- function(x,
-                          range = "default",
-                          ci = 0.95,
-                          ci_method = "ETI",
-                          verbose = TRUE,
-                          ...) {
+rope.MCMCglmm <- function(
+  x,
+  range = "default",
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  verbose = TRUE,
+  ...
+) {
   nF <- x$Fixed$nfl
   out <- rope(
     as.data.frame(x$Sol[, 1:nF, drop = FALSE]),
     range = range,
     ci = ci,
     ci_method = ci_method,
+    complement = complement,
     verbose = verbose,
     ...
   )
@@ -408,17 +462,21 @@ rope.MCMCglmm <- function(x,
 
 
 #' @export
-rope.mcmc <- function(x,
-                      range = "default",
-                      ci = 0.95,
-                      ci_method = "ETI",
-                      verbose = TRUE,
-                      ...) {
+rope.mcmc <- function(
+  x,
+  range = "default",
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  verbose = TRUE,
+  ...
+) {
   out <- rope(
     as.data.frame(x),
     range = range,
     ci = ci,
     ci_method = ci_method,
+    complement = complement,
     verbose = verbose,
     ...
   )
@@ -429,13 +487,24 @@ rope.mcmc <- function(x,
 
 
 #' @export
-rope.bcplm <- function(x,
-                       range = "default",
-                       ci = 0.95,
-                       ci_method = "ETI",
-                       verbose = TRUE,
-                       ...) {
-  out <- rope(insight::get_parameters(x), range = range, ci = ci, ci_method = ci_method, verbose = verbose, ...)
+rope.bcplm <- function(
+  x,
+  range = "default",
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  verbose = TRUE,
+  ...
+) {
+  out <- rope(
+    insight::get_parameters(x, verbose = verbose),
+    range = range,
+    ci = ci,
+    ci_method = ci_method,
+    complement = complement,
+    verbose = verbose,
+    ...
+  )
   attr(out, "object_name") <- NULL
   attr(out, "data") <- insight::safe_deparse_symbol(substitute(x))
   out
@@ -454,30 +523,45 @@ rope.BGGM <- rope.bcplm
 rope.mcmc.list <- rope.bcplm
 
 
+#' @rdname rope
 #' @export
-rope.stanreg <- function(x,
-                         range = "default",
-                         ci = 0.95,
-                         ci_method = "ETI",
-                         effects = "fixed",
-                         component = "location",
-                         parameters = NULL,
-                         verbose = TRUE,
-                         ...) {
+rope.stanreg <- function(
+  x,
+  range = "default",
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  effects = "fixed",
+  component = "location",
+  parameters = NULL,
+  verbose = TRUE,
+  ...
+) {
   if (all(range == "default")) {
     range <- rope_range(x, verbose = verbose)
   } else if (!is.list(range) && (!all(is.numeric(range)) || length(range) != 2)) {
-    insight::format_error("`range` should be 'default' or a vector of 2 numeric values (e.g., c(-0.1, 0.1)).")
+    insight::format_error(
+      "`range` should be 'default' or a vector of 2 numeric values (e.g., c(-0.1, 0.1))."
+    )
   }
 
   # check for possible collinearity that might bias ROPE
-  if (verbose && !inherits(x, "blavaan")) .check_multicollinearity(x, "rope")
+  if (verbose && !inherits(x, "blavaan")) {
+    .check_multicollinearity(x, "rope")
+  }
 
   rope_data <- rope(
-    insight::get_parameters(x, effects = effects, component = component, parameters = parameters),
+    insight::get_parameters(
+      x,
+      effects = effects,
+      component = component,
+      parameters = parameters,
+      verbose = verbose
+    ),
     range = range,
     ci = ci,
     ci_method = ci_method,
+    complement = complement,
     verbose = verbose,
     ...
   )
@@ -504,15 +588,18 @@ rope.blavaan <- rope.stanreg
 
 #' @rdname rope
 #' @export
-rope.brmsfit <- function(x,
-                         range = "default",
-                         ci = 0.95,
-                         ci_method = "ETI",
-                         effects = "fixed",
-                         component = "conditional",
-                         parameters = NULL,
-                         verbose = TRUE,
-                         ...) {
+rope.brmsfit <- function(
+  x,
+  range = "default",
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  effects = "fixed",
+  component = "conditional",
+  parameters = NULL,
+  verbose = TRUE,
+  ...
+) {
   # check range argument
   if (all(range == "default")) {
     range <- rope_range(x, verbose = verbose)
@@ -535,7 +622,9 @@ rope.brmsfit <- function(x,
   }
 
   # check for possible collinearity that might bias ROPE and print a warning
-  if (verbose) .check_multicollinearity(x, "rope")
+  if (verbose) {
+    .check_multicollinearity(x, "rope")
+  }
 
   # calc rope
   if (insight::is_multivariate(x)) {
@@ -552,11 +641,13 @@ rope.brmsfit <- function(x,
             effects = effects,
             component = component,
             parameters = parameters,
+            verbose = verbose,
             ...
           ),
           range = range[[dv_item]],
           ci = ci,
           ci_method = ci_method,
+          complement = complement,
           verbose = verbose,
           ...
         )
@@ -578,10 +669,17 @@ rope.brmsfit <- function(x,
     )
   } else {
     rope_data <- rope(
-      insight::get_parameters(x, effects = effects, component = component, parameters = parameters),
+      insight::get_parameters(
+        x,
+        effects = effects,
+        component = component,
+        parameters = parameters,
+        verbose = verbose
+      ),
       range = range,
       ci = ci,
       ci_method = ci_method,
+      complement = complement,
       verbose = verbose,
       ...
     )
@@ -598,24 +696,34 @@ rope.brmsfit <- function(x,
 
 
 #' @export
-rope.sim.merMod <- function(x,
-                            range = "default",
-                            ci = 0.95,
-                            ci_method = "ETI",
-                            effects = "fixed",
-                            parameters = NULL,
-                            verbose = TRUE,
-                            ...) {
+rope.sim.merMod <- function(
+  x,
+  range = "default",
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  effects = "fixed",
+  parameters = NULL,
+  verbose = TRUE,
+  ...
+) {
   if (all(range == "default")) {
     range <- rope_range(x, verbose = verbose)
   } else if (!is.list(range) && (!all(is.numeric(range)) || length(range) != 2)) {
-    insight::format_error("`range` should be 'default' or a vector of 2 numeric values (e.g., c(-0.1, 0.1)).")
+    insight::format_error(
+      "`range` should be 'default' or a vector of 2 numeric values (e.g., c(-0.1, 0.1))."
+    )
   }
 
   rope_list <- lapply(c("fixed", "random"), function(.x) {
-    parms <- insight::get_parameters(x, effects = .x, parameters = parameters)
+    parms <- insight::get_parameters(
+      x,
+      effects = .x,
+      parameters = parameters,
+      verbose = verbose
+    )
 
-    getropedata <- .prepare_rope_df(parms, range, ci, ci_method, verbose)
+    getropedata <- .prepare_rope_df(parms, range, ci, ci_method, complement, verbose)
     tmp <- getropedata$tmp
     HDI_area <- getropedata$HDI_area
 
@@ -642,7 +750,8 @@ rope.sim.merMod <- function(x,
     args = c(insight::compact_list(rope_list), make.row.names = FALSE)
   )
 
-  dat <- switch(effects,
+  dat <- switch(
+    effects,
     fixed = .select_rows(dat, "Group", "fixed"),
     random = .select_rows(dat, "Group", "random"),
     dat
@@ -668,21 +777,26 @@ rope.sim.merMod <- function(x,
 
 
 #' @export
-rope.sim <- function(x,
-                     range = "default",
-                     ci = 0.95,
-                     ci_method = "ETI",
-                     parameters = NULL,
-                     verbose = TRUE,
-                     ...) {
+rope.sim <- function(
+  x,
+  range = "default",
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  parameters = NULL,
+  verbose = TRUE,
+  ...
+) {
   if (all(range == "default")) {
     range <- rope_range(x, verbose = verbose)
   } else if (!is.list(range) && (!all(is.numeric(range)) || length(range) != 2)) {
-    insight::format_error("`range` should be 'default' or a vector of 2 numeric values (e.g., c(-0.1, 0.1)).")
+    insight::format_error(
+      "`range` should be 'default' or a vector of 2 numeric values (e.g., c(-0.1, 0.1))."
+    )
   }
 
-  parms <- insight::get_parameters(x, parameters = parameters)
-  getropedata <- .prepare_rope_df(parms, range, ci, ci_method, verbose)
+  parms <- insight::get_parameters(x, parameters = parameters, verbose = verbose)
+  getropedata <- .prepare_rope_df(parms, range, ci, ci_method, complement, verbose)
 
   dat <- getropedata$tmp
   HDI_area <- getropedata$HDI_area
@@ -710,19 +824,25 @@ rope.sim <- function(x,
 
 # helper -------------------------------------------------------------------
 
-
 #' @keywords internal
-.rope <- function(x, range = c(-0.1, 0.1), ci = 0.95, ci_method = "ETI", verbose = TRUE) {
+.rope <- function(
+  x,
+  range = c(-0.1, 0.1),
+  ci = 0.95,
+  ci_method = "ETI",
+  complement = FALSE,
+  verbose = TRUE
+) {
   ci_bounds <- ci(x, ci = ci, method = ci_method, verbose = verbose)
 
   if (anyNA(ci_bounds)) {
-    rope_percentage <- NA
+    inferiority_percentage <- superiority_percentage <- rope_percentage <- NA
   } else {
     HDI_area <- x[x >= ci_bounds$CI_low & x <= ci_bounds$CI_high]
-    area_within <- HDI_area[HDI_area >= min(range) & HDI_area <= max(range)]
-    rope_percentage <- length(area_within) / length(HDI_area)
+    rope_percentage <- mean(HDI_area >= min(range) & HDI_area <= max(range))
+    superiority_percentage <- mean(HDI_area > max(range))
+    inferiority_percentage <- mean(HDI_area < min(range))
   }
-
 
   rope <- data.frame(
     CI = ci,
@@ -730,6 +850,11 @@ rope.sim <- function(x,
     ROPE_high = range[2],
     ROPE_Percentage = rope_percentage
   )
+
+  if (isTRUE(complement)) {
+    rope[["Superiority_Percentage"]] <- superiority_percentage
+    rope[["Inferiority_Percentage"]] <- inferiority_percentage
+  }
 
   attr(rope, "HDI_area") <- c(ci_bounds$CI_low, ci_bounds$CI_high)
   attr(rope, "CI_bounds") <- c(ci_bounds$CI_low, ci_bounds$CI_high)
@@ -739,7 +864,7 @@ rope.sim <- function(x,
 
 
 #' @keywords internal
-.prepare_rope_df <- function(parms, range, ci, ci_method, verbose) {
+.prepare_rope_df <- function(parms, range, ci, ci_method, complement, verbose) {
   if (is.list(range)) {
     # check if list of values contains only valid values
     range <- .check_list_range(range, parms)
@@ -751,6 +876,7 @@ rope.sim <- function(x,
           range = r,
           ci = ci,
           ci_method = ci_method,
+          complement = complement,
           verbose = verbose
         )
       },
@@ -765,6 +891,7 @@ rope.sim <- function(x,
       range = range,
       ci = ci,
       ci_method = ci_method,
+      complement = complement,
       verbose = verbose,
       simplify = FALSE
     )